# api/docs/swagger.yaml
definitions:
...
  editory_submission_api_models.CreateEditorReq:
    properties:
      email:
        type: string
      first_name:
        type: string
      journal_id:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  editory_submission_api_models.CreateEditorRes:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/editory_submission_api_models.Role'
    type: object
...
  editory_submission_genproto_auth_service.GetUserListRes:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/editory_submission_genproto_auth_service.User'
        type: array
    type: object
...
  editory_submission_genproto_auth_service.User:
    properties:
      address:
        type: string
      city_id:
        type: string
      country_id:
        type: string
      degree:
        type: string
      email:
        type: string
      email_verification:
        type: boolean
      extra_phone:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      post_code:
        type: string
      prof_sphere:
        type: string
      role:
        items:
          $ref: '#/definitions/editory_submission_genproto_auth_service.Role'
        type: array
      username:
        type: string
    type: object
...
  editory_submission_genproto_content_service.GetJournalListRes:
    properties:
      count:
        type: integer
      journals:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
        type: array
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
  /admin/author:
    get:
      consumes:
      - application/json
      description: Get Author List
      operationId: get_admin_author_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetEditorListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.GetUserListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Author List
      tags:
      - Admin
  /admin/editor:
    get:
      consumes:
      - application/json
      description: Get Editor List
      operationId: get_editor_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: journal-id
        in: query
        name: journal-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetEditorListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.GetUserListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Editor List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create Editor
      operationId: create_admin_editor
      parameters:
      - description: CreateEditorRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/editory_submission_api_models.CreateEditorReq'
      produces:
      - application/json
      responses:
        "201":
          description: Editor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_api_models.CreateEditorRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Editor
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Editor
      operationId: update_editor
      parameters:
      - description: UpdateUserRequestBody
        in: body
        name: editor
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_auth_service.User'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Editor
      tags:
      - Admin
...
  /admin/journal:
    get:
      consumes:
      - application/json
      description: Get Journal List
      operationId: get_admin_journal_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetJournalListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetJournalListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Journal List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 'Create Journal types: [''EDITOR_SPOTLIGHT'', ''SPECIAL_ISSUE'',
        ''ABOUT_JOURNAL'', ''EDITORIAL_BARD'', ''PEER_REVIEW_PROCESS'', ''PUBLICATION_ETHICS'',
        ''ABSTRACTING_INDEXING'', ''ARTICLE_PROCESSING_CHARGES'']'
      operationId: create_admin_journal
      parameters:
      - description: CreateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.CreateJournalReq'
      produces:
      - application/json
      responses:
        "201":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Journal
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Journal
      operationId: update_admin_journal
      parameters:
      - description: UpdateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
      produces:
      - application/json
      responses:
        "200":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Journal
      tags:
      - Admin
...
  /admin/subject:
    get:
      consumes:
      - application/json
      description: Get Subject List
      operationId: get_admin_subject_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetSubjectListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetSubjectListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Subject List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create Subject
      operationId: create_subject
      parameters:
      - description: CreateSubjectRequestBody
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.CreateSubjectReq'
      produces:
      - application/json
      responses:
        "201":
          description: Subject data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.CreateSubjectRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Subject
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Subject
      operationId: update_subject
      parameters:
      - description: UpdateSubjectRequestBody
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.UpdateSubjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: Subject data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.UpdateSubjectRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Subject
      tags:
      - Admin
...
  /admin/university:
    get:
      consumes:
      - application/json
      description: Get University List
      operationId: get_admin_university_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUniversityListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetUniversityListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get University List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create University
      operationId: create_university
      parameters:
      - description: CreateUniversityRequestBody
        in: body
        name: university
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.CreateUniversityReq'
      produces:
      - application/json
      responses:
        "201":
          description: University data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.CreateUniversityRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create University
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update University
      operationId: update_university
      parameters:
      - description: UpdateUniversityRequestBody
        in: body
        name: university
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.UpdateUniversityReq'
      produces:
      - application/json
      responses:
        "200":
          description: University data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.UpdateUniversityRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update University
      tags:
      - Admin
...
  /city:
    get:
      consumes:
      - application/json
      description: Get City List
      operationId: get_city_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: country-id
        in: query
        name: country-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCityListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetCityListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get City List
      tags:
      - General
  /config:
    get:
      consumes:
      - application/json
      description: show service config data when the service environment set to debug
        mode
      operationId: get_config
      produces:
      - application/json
      responses:
        "200":
          description: Response data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/config.Config'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
      summary: get config data on the debug mode
  /country:
    get:
      consumes:
      - application/json
      description: Get Country List
      operationId: get_country_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCountryListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetCountryListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Country List
      tags:
      - General
  /has-access:
    post:
      consumes:
      - application/json
      description: Has Access
      operationId: has_access
      parameters:
      - description: HasAccessRequestBody
        in: body
        name: has-access
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_auth_service.HasAccessReq'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.HasAccessRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Has Access
      tags:
      - Session
  /journal:
    put:
      consumes:
      - application/json
      description: Update Journal
      operationId: update_journal
      parameters:
      - description: UpdateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
      produces:
      - application/json
      responses:
        "200":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Journal
      tags:
      - Journal
...
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_auth_service.LoginReq'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.LoginRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Session
  /logout:
    delete:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_auth_service.LogoutReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Logout User
      tags:
      - Session
...
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_auth_service.RefreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_auth_service.RefreshTokenRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
  /send-verification-message:
    post:
      consumes:
      - application/json
      description: SendVerificationMessage
      operationId: send_verification_message
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editory_submission_api_models.SendVerificationMessageReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Send Verification Message
      tags:
      - User
  /subject:
    get:
      consumes:
      - application/json
      description: Get Subject List
      operationId: get_subject_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetSubjectListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetSubjectListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Subject List
      tags:
      - General
...
swagger: "2.0"